Uploading old class project
This is my final project for a web apps class that I took. The site is
hosted on http://surveybasket.heroku.com but I’ve included all of the
original files here including screenshots and a new README that
describes the project.

It’s not recommended that the packages and frameworks be updated since
they are a couple years out of date at this point but everything has
been confirmed to run properly (at the time of this commit) if the
README instructions are followed properly.